added run from docker hub to readme, fixed animation execution and refactored animnation routes
