+++ README.md
@@ -12,7 +12,13 @@ This project borrows heavily from the work of [Don Najd](https://github.com/dnaj
 
 ## Quick Start
 
-### 1. Build the Docker Image
+Run the latest release from docker hub
+
+```bash
+docker  run -it -p 3000:3000 -e NAO_IP=<YOUR NAO ROBOT IP> davesnowdon/nao-bridge:latest
+```
+
+## Build and run the Docker image locally
 
 Navigate to the server directory and build the Docker image:
 
@@ -21,19 +27,13 @@ cd server
 docker build -t nao-bridge .
 ```
 
-### 2. Run the Container
-
-Set your NAO robot's IP address and run the container:
+run the container:
 
 ```bash
-# Set your NAO robot's IP address
-export NAO_IP=192.168.1.100
-
-# Run the container
-docker run -p 3000:3000 -e NAO_IP=$NAO_IP nao-bridge
+docker run -p 3000:3000 -e NAO_IP=<YOUR NAO ROBOT IP> nao-bridge
 ```
 
-### 3. Verify the API is Running
+## Verify the API is Running
 
 The API server will be available at `http://localhost:3000`. You can test it with:
